+++ start/src/App.vue
@@ -1,15 +1,15 @@
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+  const goBack = (): void => {
+    console.log('go back')
+  }
+</script>
 
 <template>
-  <f-list>
-    <f-list-item>
-      这是一段文字这是一段文字这是一段文字这是一段文字
-      <f-button font-color="#fff" color="#ff0de5" round> 自定义颜色 </f-button>
-      <f-button font-color="#fff" color="#22c1c3" loading> loading </f-button>
-    </f-list-item>
-    <f-list-item>这是一段文字</f-list-item>
-    <f-list-item>这是一段文字</f-list-item>
-    <f-list-item>这是一段文字</f-list-item>
-    <f-list-item>这是一段文字</f-list-item>
-  </f-list>
+  <f-page-header
+    title="标题"
+    subtitle="副标题"
+    title-bold
+    title-color="red"
+    @back="goBack"
+  />
 </template>